parsers/handlebars: Extract common base class
