+++ src/parsers/handlebars/glimmer.js
@@ -1,4 +1,4 @@
-import defaultParserInterface from '../utils/defaultParserInterface';
+import defaultParserInterface from './utils/defaultHandlebarsParserInterface';
 import pkg from 'glimmer-syntax/package.json';
 
 const ID = 'glimmer';
@@ -11,29 +11,8 @@ export default {
   version: pkg.version,
   homepage: pkg.homepage,
 
-  locationProps: new Set(['loc']),
-
   loadParser(callback) {
-    require(['glimmer-syntax'], callback);
-  },
-
-  parse(glimmer, code, options) {
-    this.lineOffsets = [];
-    let index = 0;
-    do {
-      this.lineOffsets.push(index);
-    } while (index = code.indexOf('\n', index) + 1); // eslint-disable-line no-cond-assign
-
-    return glimmer.preprocess(code, options);
-  },
-
-  getOffset({ line, column }) {
-    return this.lineOffsets[line - 1] + column;
-  },
-
-  nodeToRange({ loc }) {
-    if (!loc) return;
-    return [loc.start, loc.end].map(pos => this.getOffset(pos));
+    require(['glimmer-syntax'], (glimmer) => callback(glimmer.preprocess));
   },
 
   opensByDefault(node, key) {
