+++ src/parsers/handlebars/handlebars.js
@@ -1,4 +1,4 @@
-import defaultParserInterface from '../utils/defaultParserInterface';
+import defaultParserInterface from './utils/defaultHandlebarsParserInterface';
 import pkg from 'handlebars/package.json';
 
 const ID = 'handlebars';
@@ -11,29 +11,8 @@ export default {
   version: pkg.version,
   homepage: pkg.homepage,
 
-  locationProps: new Set(['loc']),
-
   loadParser(callback) {
-    require(['handlebars'], callback);
-  },
-
-  parse(handlebars, code, options) {
-    this.lineOffsets = [];
-    let index = 0;
-    do {
-      this.lineOffsets.push(index);
-    } while (index = code.indexOf('\n', index) + 1); // eslint-disable-line no-cond-assign
-
-    return handlebars.parse(code, options);
-  },
-
-  getOffset({ line, column }) {
-    return this.lineOffsets[line - 1] + column;
-  },
-
-  nodeToRange({ loc }) {
-    if (!loc) return;
-    return [loc.start, loc.end].map(pos => this.getOffset(pos));
+    require(['handlebars'], (handlebars) => callback(handlebars.parse));
   },
 
   opensByDefault(node, key) {
